fix(F0Callout): split the skeleton out so its docs page can render
`CalloutInternal.tsx` exported both `CalloutInternal` and `CalloutSkeleton`,
and react-docgen 8 fails a file with two exported components: "Multiple
exported component definitions found". Storybook surfaces that as an empty
docs page, which is the page someone reads to migrate off a deprecated
component.

Verified by running react-docgen on both files: it failed before and parses
one component per file now.
