+++ packages/react/src/kits/ai/Banners/F0Callout/CalloutInternal.tsx
@@ -5,10 +5,8 @@ import { F0Icon, IconType } from "@/components/F0Icon"
 import { CheckCircle, Cross, InfoCircle, Warning } from "@/icons/app"
 import { OneEllipsis } from "@/lib/OneEllipsis"
 import { cn } from "@/lib/utils"
-import { Skeleton } from "@/ui/skeleton"
-import { CalloutInternalProps, CalloutSkeletonProps } from "./types"
-
-const calloutVariants = cva({
+import { CalloutInternalProps } from "./types"
+export const calloutVariants = cva({
   base: "flex w-full flex-col rounded-lg p-[1px]",
   variants: {
     variant: {
@@ -111,41 +109,3 @@ export const CalloutInternal = forwardRef<HTMLDivElement, CalloutInternalProps>(
     )
   }
 )
-
-export const CalloutSkeleton = ({
-  compact,
-  variant = "ai",
-}: CalloutSkeletonProps) => {
-  return (
-    <div
-      className={calloutVariants({ variant })}
-      aria-busy="true"
-      aria-live="polite"
-    >
-      <div className="flex flex-row items-center justify-between px-4 py-2">
-        <Skeleton className="h-5 w-32 rounded-md" />
-      </div>
-
-      <div className="flex flex-col gap-[1px]">
-        <div
-          className={cn(
-            "rounded-t-[13.25px] bg-f1-background px-4 py-3",
-            compact && "rounded-[13.25px]"
-          )}
-        >
-          <div className="flex flex-col gap-2">
-            <Skeleton className="h-4 w-full rounded-md" />
-            <Skeleton className="h-4 w-3/4 rounded-md" />
-            <Skeleton className="h-4 w-1/2 rounded-md" />
-          </div>
-        </div>
-        {!compact ? (
-          <div className="flex flex-row items-center justify-between gap-3 rounded-b-[13.25px] bg-f1-background px-4 py-3">
-            <Skeleton className="h-8 w-24 rounded-md" />
-            <Skeleton className="h-8 w-28 rounded-md" />
-          </div>
-        ) : null}
-      </div>
-    </div>
-  )
-}
