+++ packages/react/src/kits/ai/Banners/F0Callout/index.stories.tsx
@@ -2,9 +2,8 @@ import { StoryFn, StoryObj } from "@storybook/react-vite"
 import { expect, within } from "storybook/test"
 import { CheckDouble, ExternalLink } from "@/icons/app"
 import { WithDataTestIdProps } from "@/lib/data-testid"
-import { CalloutSkeleton } from "./CalloutInternal"
+import { CalloutSkeleton } from "./CalloutSkeleton"
 import { F0Callout, F0CalloutProps } from "."
-
 const meta = {
   title: "AI/AICallout (deprecated)",
   component: F0Callout,
