+++ packages/react/src/kits/ai/Banners/F0Callout/index.tsx
@@ -1,8 +1,8 @@
 import { forwardRef } from "react"
 import { withDataTestId } from "@/lib/data-testid"
 import { withSkeleton } from "@/lib/skeleton"
-import { CalloutInternal, CalloutSkeleton } from "./CalloutInternal"
-import { CalloutInternalProps, CalloutSkeletonProps } from "./types"
+import { CalloutInternal } from "./CalloutInternal"
+import { CalloutSkeleton } from "./CalloutSkeleton"import { CalloutInternalProps, CalloutSkeletonProps } from "./types"
 
 export type F0CalloutProps = CalloutInternalProps
